feat: Cambiar tipo de MovimientoId a long en DTOs y modelos para alinearse con SQL Server y evitar suposiciones de int32

+++ GesvenApi/Models/Dtos/Responses/MovimientoDto.cs
@@ -5,7 +5,7 @@ namespace GesvenApi.Models.Dtos.Responses;
 /// </summary>
 public class MovimientoDto
 {
-    public int MovimientoId { get; set; }
+    public long MovimientoId { get; set; }
     public int InstalacionId { get; set; }
     public int ProductoId { get; set; }
     public string ProductoNombre { get; set; } = string.Empty;
